Conflicts:
	RapidFTR-Android/src/main/java/com/rapidftr/activity/RapidFtrActivity.java
	RapidFTR-Android/src/main/java/com/rapidftr/activity/ViewChildActivity.java
	RapidFTR-Android/src/main/java/com/rapidftr/service/ChildService.java
	RapidFTR-Android/src/test/java/com/rapidftr/service/ChildServiceTest.java

+++ RapidFTR-Android/src/main/java/com/rapidftr/utils/AudioCaptureHelper.java
@@ -14,7 +14,7 @@ public class AudioCaptureHelper extends CaptureHelper{
 
     public void saveAudio(Child child, InputStream inputStream) throws JSONException, IOException {
         File file = new File(getDir(), child.optString("recorded_audio"));
-        if (!file.exists()){
+        if (!file.exists() && !child.optString("recorded_audio").equals("")){
             IOUtils.copy(inputStream, new FileOutputStream(file));
         }
     }
